Moved data.js file to be nested in form folder and update filepaths
